fix: fix support

+++ src/support/message-tiket/message-tiket.service.ts
@@ -1,6 +1,7 @@
 import { Injectable } from '@nestjs/common'
 import { PrismaService } from '../../database/prisma.service'
 import { CreateMessageTiketInput } from './dto/create-message-tiket.input'
+import { FindAllMessageTiket } from './dto/find-all-message-tiket'
 import { UpdateMessageTiketInput } from './dto/update-message-tiket.input'
 import { MessageTiket } from './entities/message-tiket.entity'
 
@@ -24,8 +25,9 @@ export class MessageTiketService {
         })
     }
 
-    async findAllMessageTiket(): Promise<MessageTiket[]> {
+    async findAllMessageTiket(findAllMessageTiket: FindAllMessageTiket): Promise<MessageTiket[]> {
         return await this.prisma.messageTiket.findMany({
+            where: findAllMessageTiket,
             include: {
                 toWhom: true,
                 fromWhom: true,
@@ -34,24 +36,45 @@ export class MessageTiketService {
     }
 
     async findOneMessageTiket(tiketId: number): Promise<MessageTiket> {
-        return await this.prisma.messageTiket.findFirst({
+        const message = await this.prisma.messageTiket.findFirst({
             where: { id: tiketId },
             include: {
                 toWhom: true,
                 fromWhom: true,
             },
         })
+
+        if (!message) {
+            throw new Error('Message is not exist')
+        }
+        return message
     }
 
     async updateMessageTiket(tiketId: number, updateMessageTiketInput: UpdateMessageTiketInput): Promise<MessageTiket> {
-        return await this.prisma.messageTiket.update({
+        const message = await this.prisma.messageTiket.update({
             where: { id: tiketId },
             data: updateMessageTiketInput,
             include: { toWhom: true, fromWhom: true },
         })
+
+        if (!message) {
+            throw new Error('Message is not exist')
+        }
+        return message
     }
 
     async removeMessageTiket(tiketId: number): Promise<MessageTiket> {
+        const message = await this.prisma.messageTiket.findFirst({
+            where: { id: tiketId },
+            include: {
+                toWhom: true,
+                fromWhom: true,
+            },
+        })
+
+        if (!message) {
+            throw new Error('Message is not exist')
+        }
         return await this.prisma.messageTiket.delete({
             where: { id: tiketId },
             include: {
